fix: Fix bug in tag page
Switch tab button does not display the number of readings.
