+++ components/PageInfo.vue
@@ -86,8 +86,7 @@ export default {
   },
   methods: {
     goTags (tag) {
-      const base = this.$site.base
-      window.location.href = `${base}tag/?tag=${tag}`
+      this.$router.push({ path: `/tags/${tag}/` })
     }
   }
 }
